+++ src/components/MapList.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import React from "react";
 import MapListCard from "./MapListCard";
 import { SchoolMapPin } from "@/types/school";
@@ -28,7 +30,7 @@ const MapList = ({
 }: MapListProps) => {
   return (
     <div className="flex flex-col overflow-auto">
-      <div className="flex flex-col overflow-auto gap-2 max-md:mb-4 md:gap-4">
+      <div className="flex flex-col gap-2 overflow-auto max-md:mb-4 md:gap-4">
         {schools
           .sort((a, b) => a.name.localeCompare(b.name))
           .map((school) => (
